Refactor InnerTicketInfo
